+++ tests/ui/widgets/test_systray.py
@@ -3,7 +3,7 @@ from gi.repository import Gtk
 from wiring.scanning import scan_to_graph
 
 from tests.conftest import refresh_gui
-from tomate.pomodoro.event import Bus, Events, connect_events
+from tomate.pomodoro.event import Events
 from tomate.ui.widgets import TrayIconMenu
 
 
@@ -13,15 +13,11 @@ def view():
 
 
 @pytest.fixture
-def subject(graph, view):
-    Bus.receivers.clear()
-
+def subject(graph, bus, view):
+    graph.register_instance("tomate.bus", bus)
     graph.register_instance("tomate.ui.view", view)
     scan_to_graph(["tomate.ui.widgets.systray"], graph)
-    instance = graph.get("tomate.ui.tray.menu")
-
-    connect_events(instance)
-    return instance
+    return graph.get("tomate.ui.tray.menu")
 
 
 def test_module(graph, subject):
@@ -52,8 +48,8 @@ def test_show_window_when_hide_item_is_clicked(view, subject):
 
 
 @pytest.mark.parametrize("event,hide,show", [(Events.WINDOW_HIDE, False, True), (Events.WINDOW_SHOW, True, False)])
-def test_change_items_visibility(event, hide, show, subject):
-    Bus.send(event)
+def test_change_items_visibility(event, hide, show, bus, subject):
+    bus.send(event)
 
     assert subject.hide_item.get_visible() is hide
     assert subject.show_item.get_visible() is show