+++ tomate/pomodoro/session.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import enum
+import logging
 import uuid
 from collections import namedtuple
 
@@ -8,10 +9,11 @@ from blinker import Signal
 from wiring import SingletonScope, inject
 from wiring.scanning import register
 
-from .event import Bus, Events, Subscriber, on
+from .event import Events, Subscriber, on
 from .fsm import fsm
 from .timer import Payload as TimerPayload, SECONDS_IN_A_MINUTE, Timer
 
+logger = logging.getLogger(__name__)
 Payload = namedtuple("SessionPayload", "id type pomodoros duration")
 EndPayload = namedtuple("SessionEndPayload", "id type pomodoros duration previous")
 
@@ -40,20 +42,26 @@ class State(enum.Enum):
 
 @register.factory("tomate.session", scope=SingletonScope)
 class Session(Subscriber):
-    @inject(timer="tomate.timer", config="tomate.config", bus="tomate.bus")
-    def __init__(self, timer: Timer, config, bus: Signal):
+    @inject(
+        bus="tomate.bus",
+        config="tomate.config",
+        timer="tomate.timer",
+    )
+    def __init__(self, bus: Signal, config, timer: Timer):
         self._config = config
         self._timer = timer
         self._bus = bus
         self.state = State.STOPPED
         self.current = Type.POMODORO
         self.pomodoros = 0
+        self.connect(bus)
 
     @fsm(
         source=[State.STOPPED, State.ENDED], target=State.STARTED, exit=lambda self: self._trigger(Events.SESSION_START)
     )
     def start(self) -> bool:
         self._timer.start(self.duration)
+        logger.debug("action=start")
         return True
 
     def is_running(self) -> bool:
@@ -67,18 +75,32 @@ class Session(Subscriber):
     )
     def stop(self) -> bool:
         self._timer.stop()
+        logger.debug("action=stop")
         return True
 
     @fsm(source=[State.STOPPED, State.ENDED], target="self", exit=lambda self: self._trigger(Events.SESSION_RESET))
     def reset(self) -> bool:
         self.pomodoros = 0
+        logger.debug("action=reset")
         return True
+
+    @on(Events.CONFIG_CHANGE)
+    @fsm(source=[State.STOPPED, State.ENDED], target="self", exit=lambda self: self._trigger(Events.SESSION_CHANGE))
+    def change(self, *_, **kwargs) -> bool:
+        self.current = kwargs.get("session", self.current)
+        logger.debug("action=change session=%s", self.current)
+        return True
+
+    @property
+    def duration(self) -> int:
+        minutes = self._config.get_int("Timer", self.current.option)
+        return int(minutes * SECONDS_IN_A_MINUTE)
 
     def timer_is_up(self) -> bool:
         return not self._timer.is_running()
 
+    @on(Events.TIMER_END)
     @fsm(source=[State.STARTED], target=State.ENDED, condition=timer_is_up)
-    @on(Bus, [Events.TIMER_END])
     def _end(self, _, payload: TimerPayload) -> bool:
         previous = self._create_payload(duration=payload.duration)
 
@@ -88,22 +110,12 @@ class Session(Subscriber):
         else:
             self.current = Type.POMODORO
 
+        logger.debug("action=end previous=%s current=%s", previous.type, self.current)
+
         self.state = State.ENDED
-        self._trigger_end(previous)
+        self._bus.send(Events.SESSION_END, payload=self._create_payload(EndPayload, previous=previous))
 
         return True
-
-    def _trigger_end(self, previous: Payload):
-        self._bus.send(
-            Events.SESSION_END,
-            payload=EndPayload(
-                id=uuid.uuid4(),
-                duration=self.duration,
-                pomodoros=self.pomodoros,
-                type=self.current,
-                previous=previous,
-            ),
-        )
 
     def _choose_break(self):
         return Type.LONG_BREAK if self._is_long_break() else Type.SHORT_BREAK
@@ -112,26 +124,15 @@ class Session(Subscriber):
         long_break_interval = self._config.get_int("Timer", "long_break_interval")
         return not self.pomodoros % long_break_interval
 
-    @fsm(source=[State.STOPPED, State.ENDED], target="self", exit=lambda self: self._trigger(Events.SESSION_CHANGE))
-    @on(Bus, ["config.timer"])
-    def change(self, *_, **kwargs) -> bool:
-        self.current = kwargs.get("session", self.current)
-        return True
-
-    @property
-    def duration(self) -> int:
-        minutes = self._config.get_int("Timer", self.current.option)
-        return int(minutes * SECONDS_IN_A_MINUTE)
-
     def _trigger(self, event: Events) -> None:
         self._bus.send(event, payload=self._create_payload())
 
-    def _create_payload(self, **kwargs) -> Payload:
+    def _create_payload(self, factory=Payload, **kwargs):
         defaults = {
-            "id": uuid.uuid4(),
             "duration": self.duration,
+            "id": uuid.uuid4(),
             "pomodoros": self.pomodoros,
             "type": self.current,
         }
         defaults.update(kwargs)
-        return Payload(**defaults)
+        return factory(**defaults)